+++ flutter_mpv_test/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -6,15 +6,15 @@ import FlutterMacOS
 import Foundation
 
 import file_picker
-import flutter_mpv_video
 import flutter_mpv_libs_macos_video
+import flutter_mpv_video
 import package_info_plus
 import wakelock_plus
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   FilePickerPlugin.register(with: registry.registrar(forPlugin: "FilePickerPlugin"))
-  FlutterMpvVideoPlugin.register(with: registry.registrar(forPlugin: "FlutterMpvVideoPlugin"))
   FlutterMpvLibsMacosVideoPlugin.register(with: registry.registrar(forPlugin: "FlutterMpvLibsMacosVideoPlugin"))
+  FlutterMpvVideoPlugin.register(with: registry.registrar(forPlugin: "FlutterMpvVideoPlugin"))
   FPPPackageInfoPlusPlugin.register(with: registry.registrar(forPlugin: "FPPPackageInfoPlusPlugin"))
   WakelockPlusMacosPlugin.register(with: registry.registrar(forPlugin: "WakelockPlusMacosPlugin"))
 }